code/pyproject.toml
[tool.poetry]
name = "ml-lab"
version = "0.1.0"
description = ""
authors = ["Noah Kim"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
tensorflow-macos = "^2.11.0"
tensorflow-metal = "^0.7.0"


[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
